+++ packages/nx-plugin/src/executors/internal/cli.ts
@@ -1,4 +1,7 @@
-export function createCliCommand(options?: {
+import { logger } from '@nx/devkit';
+import type { ProcessConfig } from '../../internal/execute-process.js';
+
+export function createCliCommandString(options?: {
   args?: Record<string, unknown>;
   command?: string;
   bin?: string;
@@ -7,6 +10,26 @@ export function createCliCommand(options?: {
   return `npx ${bin} ${objectToCliArgs({ _: command ?? [], ...args }).join(
     ' ',
   )}`;
+}
+
+export function createCliCommandObject(options?: {
+  args?: Record<string, unknown>;
+  command?: string;
+  bin?: string;
+}): ProcessConfig {
+  const { bin = '@code-pushup/cli', command, args } = options ?? {};
+  return {
+    command: 'npx',
+    args: [bin, ...objectToCliArgs({ _: command ?? [], ...args })],
+    observer: {
+      onError: error => {
+        logger.error(error.message);
+      },
+      onStdout: data => {
+        logger.log(data);
+      },
+    },
+  };
 }
 
 type ArgumentValue = number | string | boolean | string[];